Return teh check_configuration_file back to __main__.py
