--- demisto_sdk/utils/config.py
@@ -1,34 +1,0 @@
-import os
-from configparser import ConfigParser, MissingSectionHeaderError
-
-
-def check_configuration_file(command, args):
-    config_file_path = '.demisto-sdk-conf'
-    true_synonyms = ['true', 'True', 't', '1']
-    if os.path.isfile(config_file_path):
-        try:
-            config = ConfigParser(allow_no_value=True)
-            config.read(config_file_path)
-
-            if command in config.sections():
-                for key in config[command]:
-                    if key in args:
-                        # if the key exists in the args we will run it over if it is either:
-                        # a - a flag currently not set and is defined in the conf file
-                        # b - not a flag but an arg that is currently None and there is a value for it in the conf file
-                        if args[key] is False and config[command][key] in true_synonyms:
-                            args[key] = True
-
-                        elif args[key] is None and config[command][key] is not None:
-                            args[key] = config[command][key]
-
-                    # if the key does not exist in the current args, add it
-                    else:
-                        if config[command][key] in true_synonyms:
-                            args[key] = True
-
-                        else:
-                            args[key] = config[command][key]
-
-        except MissingSectionHeaderError:
-            pass
